Finalize DB integration
